整合本地开发内容到基础框架
- 添加MNIST示例和测试代码
- 增加脚本工具和性能基准测试
- 扩展模型格式支持（ONNX、PyTorch、TensorFlow、HuggingFace）
- 完善文档和开发指南
- 修复和优化现有实现
- 排除大文件依赖（libtorch等）

+++ src/loss/huber.c
@@ -1,6 +1,6 @@
 // huber.c - Huber loss function
-// Copyright (c) 2026 Shaoning, Xiao 萧少宁
-// Licensed under the Apache License, Version 2.0
+// Copyright (c) 2026 Boat Framework Authors
+// Distributed under the MIT License
 
 #include <boat/loss.h>
 #include <boat/tensor.h>
@@ -94,7 +94,7 @@ float huber_loss_compute(boat_loss_t* loss_ptr, const void* predictions_ptr, con
 
     // Update accumulated stats
     loss->sum += sum_loss;
-    loss->count += (int)num_elements;
+    loss->count += num_elements;
 
     return huber;
 }
